Move cache and sandbox modules to user_plugins
