+++ src/pipeline/cache/__init__.py
@@ -1,8 +1,9 @@
 """Caching utilities with pluggable backends."""
 
+from user_plugins.resources.cache_backends.redis import RedisCache
+from user_plugins.resources.cache_backends.semantic import SemanticCache
+
 from .base import CacheBackend
 from .memory import InMemoryCache
-from .redis import RedisCache
-from .semantic import SemanticCache
 
 __all__ = ["CacheBackend", "InMemoryCache", "RedisCache", "SemanticCache"]
