+++ src/user_plugins/infrastructure/__init__.py
@@ -1,0 +1,1 @@
+"""Infrastructure helpers for user plugins."""
