+++ src/user_plugins/infrastructure/sandbox/__init__.py
@@ -1,0 +1,10 @@
+"""Sandbox utilities for running and auditing plugins."""
+
+from .audit import PluginAuditor
+
+try:  # optional dependency
+    from .runner import DockerSandboxRunner
+except Exception:  # pragma: no cover - missing docker or cdktf
+    DockerSandboxRunner = None  # type: ignore
+
+__all__ = ["PluginAuditor", "DockerSandboxRunner"]
