+++ src/user_plugins/resources/__init__.py
@@ -1,0 +1,1 @@
+"""Collection of user-provided resource plugins."""
