+++ src/user_plugins/resources/cache_backends/__init__.py
@@ -1,0 +1,6 @@
+"""User-provided cache backends."""
+
+from .redis import RedisCache
+from .semantic import SemanticCache
+
+__all__ = ["RedisCache", "SemanticCache"]
