+++ src/user_plugins/resources/cache_backends/semantic.py
@@ -1,0 +1,41 @@
+from __future__ import annotations
+
+from typing import Any
+
+from pipeline.cache.base import CacheBackend
+from pipeline.resources.vectorstore import VectorStoreResource
+
+
+class SemanticCache(CacheBackend):
+    """Cache that retrieves values using vector similarity."""
+
+    def __init__(self, store: VectorStoreResource, inner: CacheBackend) -> None:
+        self.store = store
+        self.inner = inner
+        self._prompt_map: dict[str, str] = {}
+
+    async def get(self, key: str) -> Any:
+        return await self.inner.get(key)
+
+    async def set(self, key: str, value: Any, ttl: int | None = None) -> None:
+        await self.inner.set(key, value, ttl)
+
+    async def clear(self) -> None:
+        await self.inner.clear()
+
+    async def get_semantic(self, prompt: str) -> Any:
+        similar = await self.store.query_similar(prompt, 1)
+        if not similar:
+            return None
+        key = self._prompt_map.get(similar[0])
+        if not key:
+            return None
+        return await self.inner.get(key)
+
+    async def set_semantic(
+        self, prompt: str, value: Any, ttl: int | None = None
+    ) -> None:
+        key = f"semantic:{len(self._prompt_map)}"
+        self._prompt_map[prompt] = key
+        await self.store.add_embedding(prompt)
+        await self.inner.set(key, value, ttl)
